repo-updater: move vcs/git/tree/ReadTree to gitserver/commands/ReadTree
