+++ internal/gitserver/commands.go
@@ -3,16 +3,27 @@ package gitserver
 import (
 	"bytes"
 	"context"
+	"encoding/hex"
 	"fmt"
 	"io"
+	"io/fs"
 	"net/mail"
 	"os"
+	stdlibpath "path"
+	"path/filepath"
 	"strconv"
 	"strings"
+	"sync"
 	"time"
 
+	"gopkg.in/src-d/go-git.v4/plumbing/format/config"
+
+	"github.com/golang/groupcache/lru"
 	"github.com/opentracing/opentracing-go/log"
+
 	"github.com/sourcegraph/go-diff/diff"
+	"github.com/sourcegraph/sourcegraph/internal/database"
+	"github.com/sourcegraph/sourcegraph/internal/vcs/util"
 
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/api"
@@ -316,4 +327,305 @@ func (c *ClientImplementor) DiffSymbols(ctx context.Context, repo api.RepoName, 
 	command := c.Command("git", "diff", "-z", "--name-status", "--no-renames", string(commitA), string(commitB))
 	command.Repo = repo
 	return command.Output(ctx)
+}
+
+// ReadDir reads the contents of the named directory at commit.
+func ReadDir(
+	ctx context.Context,
+	db database.DB,
+	checker authz.SubRepoPermissionChecker,
+	repo api.RepoName,
+	commit api.CommitID,
+	path string,
+	recurse bool,
+) ([]fs.FileInfo, error) {
+	if Mocks.ReadDir != nil {
+		return Mocks.ReadDir(commit, path, recurse)
+	}
+
+	span, ctx := ot.StartSpanFromContext(ctx, "Git: ReadDir")
+	span.SetTag("Commit", commit)
+	span.SetTag("Path", path)
+	span.SetTag("Recurse", recurse)
+	defer span.Finish()
+
+	if err := checkSpecArgSafety(string(commit)); err != nil {
+		return nil, err
+	}
+
+	if path != "" {
+		// Trailing slash is necessary to ls-tree under the dir (not just
+		// to list the dir's tree entry in its parent dir).
+		path = filepath.Clean(util.Rel(path)) + "/"
+	}
+	files, err := lsTree(ctx, db, repo, commit, path, recurse)
+
+	if err != nil || !authz.SubRepoEnabled(checker) {
+		return files, err
+	}
+
+	a := actor.FromContext(ctx)
+	filtered, filteringErr := authz.FilterActorFileInfos(ctx, checker, a, repo, files)
+	if filteringErr != nil {
+		return nil, errors.Wrap(err, "filtering paths")
+	} else {
+		return filtered, nil
+	}
+}
+
+// lsTreeRootCache caches the result of running `git ls-tree ...` on a repository's root path
+// (because non-root paths are likely to have a lower cache hit rate). It is intended to improve the
+// perceived performance of large monorepos, where the tree for a given repo+commit (usually the
+// repo's latest commit on default branch) will be requested frequently and would take multiple
+// seconds to compute if uncached.
+var (
+	lsTreeRootCacheMu sync.Mutex
+	lsTreeRootCache   = lru.New(5)
+)
+
+// lsTree returns ls of tree at path.
+func lsTree(
+	ctx context.Context,
+	db database.DB,
+	repo api.RepoName,
+	commit api.CommitID,
+	path string,
+	recurse bool,
+) (files []fs.FileInfo, err error) {
+	if path != "" || !recurse {
+		// Only cache the root recursive ls-tree.
+		return lsTreeUncached(ctx, db, repo, commit, path, recurse)
+	}
+
+	key := string(repo) + ":" + string(commit) + ":" + path
+	lsTreeRootCacheMu.Lock()
+	v, ok := lsTreeRootCache.Get(key)
+	lsTreeRootCacheMu.Unlock()
+	var entries []fs.FileInfo
+	if ok {
+		// Cache hit.
+		entries = v.([]fs.FileInfo)
+	} else {
+		// Cache miss.
+		var err error
+		start := time.Now()
+		entries, err = lsTreeUncached(ctx, db, repo, commit, path, recurse)
+		if err != nil {
+			return nil, err
+		}
+
+		// It's only worthwhile to cache if the operation took a while and returned a lot of
+		// data. This is a heuristic.
+		if time.Since(start) > 500*time.Millisecond && len(entries) > 5000 {
+			lsTreeRootCacheMu.Lock()
+			lsTreeRootCache.Add(key, entries)
+			lsTreeRootCacheMu.Unlock()
+		}
+	}
+	return entries, nil
+}
+
+type objectInfo gitdomain.OID
+
+func (oid objectInfo) OID() gitdomain.OID { return gitdomain.OID(oid) }
+
+// LStat returns a FileInfo describing the named file at commit. If the file is a symbolic link, the
+// returned FileInfo describes the symbolic link.  lStat makes no attempt to follow the link.
+// TODO(sashaostrikov): make private when git.Stat is moved here as well
+func LStat(ctx context.Context, db database.DB, checker authz.SubRepoPermissionChecker, repo api.RepoName, commit api.CommitID, path string) (fs.FileInfo, error) {
+	span, ctx := ot.StartSpanFromContext(ctx, "Git: lStat")
+	span.SetTag("Commit", commit)
+	span.SetTag("Path", path)
+	defer span.Finish()
+
+	if err := checkSpecArgSafety(string(commit)); err != nil {
+		return nil, err
+	}
+
+	path = filepath.Clean(util.Rel(path))
+
+	if path == "." {
+		// Special case root, which is not returned by `git ls-tree`.
+		obj, err := NewClient(db).GetObject(ctx, repo, string(commit)+"^{tree}")
+		if err != nil {
+			return nil, err
+		}
+		return &util.FileInfo{Mode_: os.ModeDir, Sys_: objectInfo(obj.ID)}, nil
+	}
+
+	fis, err := lsTree(ctx, db, repo, commit, path, false)
+	if err != nil {
+		return nil, err
+	}
+	if len(fis) == 0 {
+		return nil, &os.PathError{Op: "ls-tree", Path: path, Err: os.ErrNotExist}
+	}
+
+	if !authz.SubRepoEnabled(checker) {
+		return fis[0], nil
+	}
+	// Applying sub-repo permissions
+	a := actor.FromContext(ctx)
+	include, filteringErr := authz.FilterActorFileInfo(ctx, checker, a, repo, fis[0])
+	if include && filteringErr == nil {
+		return fis[0], nil
+	} else {
+		if filteringErr != nil {
+			err = errors.Wrap(err, "filtering paths")
+		} else {
+			err = &os.PathError{Op: "ls-tree", Path: path, Err: os.ErrNotExist}
+		}
+		return nil, err
+	}
+}
+
+func lsTreeUncached(ctx context.Context, db database.DB, repo api.RepoName, commit api.CommitID, path string, recurse bool) ([]fs.FileInfo, error) {
+	if err := gitdomain.EnsureAbsoluteCommit(commit); err != nil {
+		return nil, err
+	}
+
+	// Don't call filepath.Clean(path) because ReadDir needs to pass
+	// path with a trailing slash.
+
+	if err := checkSpecArgSafety(path); err != nil {
+		return nil, err
+	}
+
+	args := []string{
+		"ls-tree",
+		"--long", // show size
+		"--full-name",
+		"-z",
+		string(commit),
+	}
+	if recurse {
+		args = append(args, "-r", "-t")
+	}
+	if path != "" {
+		args = append(args, "--", filepath.ToSlash(path))
+	}
+	cmd := NewClient(db).Command("git", args...)
+	cmd.Repo = repo
+	out, err := cmd.CombinedOutput(ctx)
+	if err != nil {
+		if bytes.Contains(out, []byte("exists on disk, but not in")) {
+			return nil, &os.PathError{Op: "ls-tree", Path: filepath.ToSlash(path), Err: os.ErrNotExist}
+		}
+		return nil, errors.WithMessage(err, fmt.Sprintf("git command %v failed (output: %q)", cmd.Args, out))
+	}
+
+	if len(out) == 0 {
+		// If we are listing the empty root tree, we will have no output.
+		if stdlibpath.Clean(path) == "." {
+			return []fs.FileInfo{}, nil
+		}
+		return nil, &os.PathError{Op: "git ls-tree", Path: path, Err: os.ErrNotExist}
+	}
+
+	trimPath := strings.TrimPrefix(path, "./")
+	lines := strings.Split(string(out), "\x00")
+	fis := make([]fs.FileInfo, len(lines)-1)
+	for i, line := range lines {
+		if i == len(lines)-1 {
+			// last entry is empty
+			continue
+		}
+
+		tabPos := strings.IndexByte(line, '\t')
+		if tabPos == -1 {
+			return nil, errors.Errorf("invalid `git ls-tree` output: %q", out)
+		}
+		info := strings.SplitN(line[:tabPos], " ", 4)
+		name := line[tabPos+1:]
+		if len(name) < len(trimPath) {
+			// This is in a submodule; return the original path to avoid a slice out of bounds panic
+			// when setting the FileInfo._Name below.
+			name = trimPath
+		}
+
+		if len(info) != 4 {
+			return nil, errors.Errorf("invalid `git ls-tree` output: %q", out)
+		}
+		typ := info[1]
+		sha := info[2]
+		if !gitdomain.IsAbsoluteRevision(sha) {
+			return nil, errors.Errorf("invalid `git ls-tree` SHA output: %q", sha)
+		}
+		oid, err := decodeOID(sha)
+		if err != nil {
+			return nil, err
+		}
+
+		sizeStr := strings.TrimSpace(info[3])
+		var size int64
+		if sizeStr != "-" {
+			// Size of "-" indicates a dir or submodule.
+			size, err = strconv.ParseInt(sizeStr, 10, 64)
+			if err != nil || size < 0 {
+				return nil, errors.Errorf("invalid `git ls-tree` size output: %q (error: %s)", sizeStr, err)
+			}
+		}
+
+		var sys interface{}
+		modeVal, err := strconv.ParseInt(info[0], 8, 32)
+		if err != nil {
+			return nil, err
+		}
+		mode := os.FileMode(modeVal)
+		switch typ {
+		case "blob":
+			const gitModeSymlink = 020000
+			if mode&gitModeSymlink != 0 {
+				mode = os.ModeSymlink
+			} else {
+				// Regular file.
+				mode = mode | 0644
+			}
+		case "commit":
+			mode = mode | gitdomain.ModeSubmodule
+			cmd := NewClient(db).Command("git", "show", fmt.Sprintf("%s:.gitmodules", commit))
+			cmd.Repo = repo
+			var submodule gitdomain.Submodule
+			if out, err := cmd.Output(ctx); err == nil {
+
+				var cfg config.Config
+				err := config.NewDecoder(bytes.NewBuffer(out)).Decode(&cfg)
+				if err != nil {
+					return nil, errors.Errorf("error parsing .gitmodules: %s", err)
+				}
+
+				submodule.Path = cfg.Section("submodule").Subsection(name).Option("path")
+				submodule.URL = cfg.Section("submodule").Subsection(name).Option("url")
+			}
+			submodule.CommitID = api.CommitID(oid.String())
+			sys = submodule
+		case "tree":
+			mode = mode | os.ModeDir
+		}
+
+		if sys == nil {
+			// Some callers might find it useful to know the object's OID.
+			sys = objectInfo(oid)
+		}
+
+		fis[i] = &util.FileInfo{
+			Name_: name, // full path relative to root (not just basename)
+			Mode_: mode,
+			Size_: size,
+			Sys_:  sys,
+		}
+	}
+	util.SortFileInfosByName(fis)
+
+	return fis, nil
+}
+
+func decodeOID(sha string) (gitdomain.OID, error) {
+	oidBytes, err := hex.DecodeString(sha)
+	if err != nil {
+		return gitdomain.OID{}, err
+	}
+	var oid gitdomain.OID
+	copy(oid[:], oidBytes)
+	return oid, nil
 }
