+++ internal/gitserver/gitdomain/common.go
@@ -3,10 +3,12 @@ package gitdomain
 import (
 	"encoding/hex"
 	"fmt"
+	"os"
 	"strings"
 	"time"
 
 	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
 // OID is a Git OID (40-char hex-encoded).
@@ -24,6 +26,33 @@ const (
 	ObjectTypeTree   ObjectType = "tree"
 	ObjectTypeBlob   ObjectType = "blob"
 )
+
+// ModeSubmodule is an os.FileMode mask indicating that the file is a Git submodule.
+//
+// To avoid being reported as a regular file mode by (os.FileMode).IsRegular, it sets other bits
+// (os.ModeDevice) beyond the Git "160000" commit mode bits. The choice of os.ModeDevice is
+// arbitrary.
+const ModeSubmodule = 0160000 | os.ModeDevice
+
+// Submodule holds information about a Git submodule and is
+// returned in the FileInfo's Sys field by Stat/ReadDir calls.
+type Submodule struct {
+	// URL is the submodule repository clone URL.
+	URL string
+
+	// Path is the path of the submodule relative to the repository root.
+	Path string
+
+	// CommitID is the pinned commit ID of the submodule (in the
+	// submodule repository's commit ID space).
+	CommitID api.CommitID
+}
+
+// ObjectInfo holds information about a Git object and is returned in (fs.FileInfo).Sys for blobs
+// and trees from Stat/ReadDir calls.
+type ObjectInfo interface {
+	OID() OID
+}
 
 // GitObject represents a GitObject
 type GitObject struct {
@@ -47,6 +76,16 @@ func IsAbsoluteRevision(s string) bool {
 		}
 	}
 	return true
+}
+
+func EnsureAbsoluteCommit(commitID api.CommitID) error {
+	// We don't want to even be running commands on non-absolute
+	// commit IDs if we can avoid it, because we can't cache the
+	// expensive part of those computations.
+	if !IsAbsoluteRevision(string(commitID)) {
+		return errors.Errorf("non-absolute commit ID: %q", commitID)
+	}
+	return nil
 }
 
 // Commit represents a git commit
